+++ test/public/home/overview.test.js
@@ -51,7 +51,7 @@ module.exports = () => {
         firstRowId = await getFirstRow(table, page);
 
         // We expect to find a table
-        expect(firstRowId).to.equal('row132');
+        expect(firstRowId).to.equal('row133');
     });
 
     it('shows correct datatypes in respective columns', async () => {