+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. See [standa
   * Tags with a length of 2 characters are now allowed
   * A notification is sent any time a log is created, not only when a log is created from the log creation page (for example auto-generated logs)
   * Main links now have a complete link behavior, such as ctrl+click to open in a new tab
+  * Changing run quality will create a log
 * Notable changes for developers:
   * Any log creation using `CreateLogUseCase` will send a notification, not only logs created from logs controller